Rename orchestrator and activity wrapper
