+++ sdk-workflows/src/test/java/io/dapr/workflows/runtime/WorkflowActivityWrapperTest.java
@@ -10,7 +10,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 
-public class ActivityWrapperTest {
+public class WorkflowActivityWrapperTest {
   public static class TestActivity implements WorkflowActivity {
     @Override
     public Object run(WorkflowActivityContext ctx) {
@@ -21,10 +21,10 @@ public class ActivityWrapperTest {
 
   @Test
   public void getName() throws NoSuchMethodException {
-    ActivityWrapper<ActivityWrapperTest.TestActivity> wrapper = new ActivityWrapper<>(
-        ActivityWrapperTest.TestActivity.class);
+    WorkflowActivityWrapper<TestActivity> wrapper = new WorkflowActivityWrapper<>(
+        WorkflowActivityWrapperTest.TestActivity.class);
     Assert.assertEquals(
-        "io.dapr.workflows.runtime.ActivityWrapperTest.TestActivity",
+        "io.dapr.workflows.runtime.WorkflowActivityWrapperTest.TestActivity",
         wrapper.getName()
     );
   }
@@ -32,8 +32,8 @@ public class ActivityWrapperTest {
   @Test
   public void createWithClass() throws NoSuchMethodException {
     TaskActivityContext mockContext = mock(TaskActivityContext.class);
-    ActivityWrapper<ActivityWrapperTest.TestActivity> wrapper = new ActivityWrapper<>(
-        ActivityWrapperTest.TestActivity.class);
+    WorkflowActivityWrapper<TestActivity> wrapper = new WorkflowActivityWrapper<>(
+        WorkflowActivityWrapperTest.TestActivity.class);
     when(mockContext.getInput(String.class)).thenReturn("Hello");
     when(mockContext.getName()).thenReturn("TestActivityContext");
     Object result = wrapper.create().run(mockContext);
